interactions: add methods to update interaction status
If a message is outgoing, update the status from daemon's signals. If a
message is incoming, the client have the ability to change the status of
the interaction from UNREAD to READ.

+++ src/api/conversation.h
@@ -40,8 +40,8 @@ struct Info
     std::string accountId;
     std::vector<std::string> participants;
     std::string callId;
-    std::map<int, interaction::Info> interactions;
-    unsigned int lastMessageUid = 0;
+    std::map<uint64_t, interaction::Info> interactions;
+    uint64_t lastMessageUid = 0;
     unsigned int unreadMessages = 0;
 };
